feat: use approval tests to verify static equality (#7293)
* feat: use approval tests to verify static equality

This is a thing that was driving me up the wall, we unnecessarily
maintain really long objects inside of our test code, while the
only thing that we *really* care about is detecting change. This
change makes writing change detectors tests super trivial. The
library used is well maintained and has been in use for a while now,
the testing pattern is super common in React land. Let me know what
you think!

* feat: add a junit rule to explicitly set the tz for test

This helps us deal with tests failing due to timezones being
different for different engineers. Now that the snapshots
take the literal TZ that was set whenever they were generated
we ned to be cognizant of that.