feat: add a limit on the total number of stores
Adds a config to limit the total number of state stores. To implement
this check, this patch converts KafkaStreamsQueryValidator into a
QueryEventListener, and uses the returned topology to compute an
estimate of the number of stores by multiplying the stores in the
topology by the number of partitions. It's convenient to implement
this as a QueryEventListener because we need to inspect the built
topology to determine the number of stores KS creates.