feat: add methods and classes to execute low level HTTP requests in the ksqldb-api-client (#8118)
* Add classes for low level HTTP calls

This PR adds methods and classes that allow users to directly execute HTTP requests against ksqldb's REST server. The API client today uses a number of "high level" classes. With an 'escape hatch' as implemented in this PR, users do not need to wait for all the REST APIs, request and response objects to be implemented as java objects and classes in the api-client.

This PR resolves one of the issues described in #8042.

Signed-off-by: Arjun Satish <arjun@confluent.io>