fix: Specify source type in DROP referential integrity errors (#8253)
Without this patch a DROP command for a non-existent table or
stream resulted in a generic error, e.g. "Source `BAAZ` does not
exist".

This patch specifies the source type in the error message. In the
example above, it might be "TABLE `BAAZ` does not exist" if the
command was "DROP TABLE BAAZ".