fix: use correct runtime to build copy of query's NamedTopology (#8474)
When building up a copy of a query's NamedTopology for use when restarting the runtime or validating a new query, we need to make sure to use a NamedTopologyBuilder that corresponds to the correct/new runtime. Otherwise the copied topology will still be tied to the original topology and its runtime.