fix: only use uuid as alter system command id (#8529)
The command id for the alter system query was using a uuid plus the query in string form. This was leading to the server to crash as the / is not allowed as a command id due to the following line final String[] splitOnSlash = fromString.split("/", 3);.

This changes the command id to only use a random uuid as originally planned.