fix: SandboxKafkaTopicClient should use defaul replication factor if applicable
If replication factor `-1` is used (implying to use the broker default) the
SandboxKafkaTopicClient created a mocked topic with replication factor 0
instead of using the broker default.

If multiple statement are submitted as once (ie, via a file), later
statement could inherit the incorrect replication factor and thus fail,
which failes all statements in the file.

Closes #4800