fix: make writes to the backup file atomic
this patch makes writes to the backup file atomic. Writing a record now will either
complete in full, or not at all. This is achieved by writing the record to a dirty
copy, and then renaming the dirty copy to the live backup file. To maintain the dirty
copy without creating a new copy on each write, we create a temporary hardlink to
the live copy before renaming the dirty copy, and then rename the link to the dirty
copy. Finally, the write appends the record to the dirty copy (which was the old
live copy) to keep it in sync