coerce property values to correct type
This patch changes Command to coerce overwritten properties to the right
types. Previously, if a user provided a property that deserializes to a
different type by default (e.g. a property that is declared as Short will
deserialize to Integer by default), the command would fail to execute due
to a type mismatch.