fix: Adds error handling for nested functions (#8850)
Addresses: https://github.com/confluentinc/ksql/issues/8324

This issue is about nested UDFs.  If a UDF has an error in processing, a null is returned.

In the case of `coalesce`, accepting a null value and doing something useful is possible.

Before this PR, the entire expression for a column was being evaluated in a single pass.  This means that
any exception would cause the entire expression to return null.

The general solution here is to wrap inner function calls with a "try-catch" which can return a null in the case of
an inner function throwing an exception.

Since such errors may represent an error in input data, the error is logged.  This handling is similar to struct and map handling.