fix: reduce query creation time with shared runtimes (#8888)
Benchmarks showed that creating 100 queries took up to 5 hours. This is likely because we currently wait on the future returned by #addNamedTopology, which blocks until all threads in the runtimes have ack'ed the new topology. Since this requires exiting the processing loop to check for topology updates before polling, and this depends on the configured max.poll.interval.ms for which we increased the default value with shared runtimes, waiting on this future can take a long time and will only increase as more queries are added.

Since we only really care about the topology being synced to the runtime's TopologyMetadata, and not whether the threads on a particular client have received it yet, we should just remove the .all().get() in SharedKafkaStreamsRuntime#start(query)