fix: Improved/fixed aggregate function error messages. (#8977)
If you call the `sum` function with an unsupported argument type, it
throws an error complaining about the `max` function. This is clearly a
copy & paste error. I've fixed it and amended the code to make future
copy & paste errors less likely. There's also some testing around this message.

Fixes #8976.

Co-authored-by: Kris Jenkins <kjenkins@confluent.io>