fix: change group name and extend CumulativeSum in ThroughputMetricsReporter (#9211)
This PR contains two fixes:

changes the group name for ksql's per-query topic-level metrics to ksql-query-throughput-metrics so that we can distinguish them from the Streams metrics they are derived from
have the ThroughputTotalMetric extend the Kafka metrics class CumulativeSum so that the ce-kafka metrics reporter we integrate with will know to compute the diff and only forward the throughput deltas (see here). This is a bit of a hack but it's necessary to ensure that the downstream metrics processing will accumulate the totals correctly.