feat: Adding ksqlDB Query Status metric.
The existing metric `query-status` reports the state of the Kafka Streams application.

This new metric `ksql-query-status` reports the state of the ksqlDB.

This is being added to show that the query can be paused.  (While a ksqlDB query is paused,
the underlying Kafka Stream application is still in the `RUNNING` state.)