feat: Allow alternative nullable representations also for protobuf_nosr
We previously introduced alternative nullable representations for nullable.
This change adds support for these properties also for protobuf_nosr.

In the process, we also fix the handling of `unwrapPrimitives` for
protobuf_nosr which was partially copied during the creation of the
protobuf_nosr format but did not have any effect.

Tested using QTT.