chore: fix tests for strict retention
Upstream Kafka streams was changed to enforce strict retention,
meaning that windows that are expired by the `RETENTION` clause
of the window, but are still available inside the materialized
`WindowStore` or `SessionStore` due to the segmented nature of
RocksDB, will _not_ be returned anymore.

KSQL functional tests were checking for the original behavior
of retention, where expired windows would still remain accessible
when querying the materialization of a persistent query with a
`RETENTION` clause. In this change, we update the expected behavior.

With the update to the new version of Kafka Streams, we change the
externally visible behavior of KSQL, but the new behavior should
be closer in line the expectations set by the KSQL documentation,
so this is not expected to cause problems for users.