test: wait for result to avoid failure due to race (#9798)
ApiTest#shouldHandleErrorInProcessingQuery test failed for no apparent reason. The assertion that was not satisfied reads a set of keys from a ConcurrentHashMap and verifies that that set is empty. The theory is that the assertion reads the set of keys before the key is concurrently removed from the hash map. This PR replaces the failing assertion with an assertion that retries the assertion for a certain amount of time before it fails.