test: fix flaky pause/resume integration test (#9799)
Fixes flaky pause/resume integration test. The test pauses a query and then it verifies that the REST app has a paused query and that the metrics show the query as paused. The latter verification has failed sometimes because the metrics have not been updated before the verification was called. This PR repeatedly verifies the metrics and only fails after the metrics do not show the paused state after a certain period of time.