fix: check for tombstone before forwarding to sink (#9876)
Checks if the record is a tombstone before it is forwarded to the sink topic.

Before this change, a tombstone record would cause a NullPointerException when ksql tries to extract the timestamp from the value since the value of a tombstone is null. The NullPointerException caused the following error log message:
"Error writing row with extracted timestamp: null","record":null,"cause":["java.lang.NullPointerException"]

If the record is a tombstone the stream time is used as the timestamp of the tombstone record.