fix: concurrency bug when doing high volume pull query over java client
Fix for a race inside vert.x’s `RecordParser`. Two threads are hitting
`RecordParserImpl.handleParsing`, one is entering from upstream via `RecordParser.handle`,
and one thread is coming from the downstream via `ReadStream.resume`. Once our
`PullQueryWriteStream` is half empty, we call the drain handler, which calls
`RecordParser.resume`, which synchronously starts parsing new records from the buffer.
`RecordParser.handleParsing` is not thread-safe.

The cause is that the monitor inside `PullQueryWriteStream` is insufficient to protect from races
inside the `RecordParser`. This fix uses a separate monitor (the object monitor of the HTTP response
object) to prevent races inside the `RecordParser`. To implement the fix, we define a shim wrapper
around `RecordParser`, which makes sure to acquire the monitor both on `handle` and on `resume`.

In the future, we may want to fix this by changing the way the drain handler is called.